net: Cleaned up websockets (moved op code and header etc)
