+++ src/net/ws/websocket.cpp
@@ -1,6 +1,6 @@
 // This file is a part of the IncludeOS unikernel - www.includeos.org
 //
-// Copyright 2016 Oslo and Akershus University College of Applied Sciences
+// Copyright 2016-2017 Oslo and Akershus University College of Applied Sciences
 // and Alfred Bratterud
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,143 +15,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <net/http/websocket.hpp>
+#include <net/ws/websocket.hpp>
 #include <kernel/os.hpp>
 #include <util/base64.hpp>
 #include <util/sha1.hpp>
 #include <cstdint>
-#include <net/http/ws_connector.hpp>
+#include <net/ws/connector.hpp>
 
-#define OPCODE_CONTINUE    0
-#define OPCODE_TEXT        1
-#define OPCODE_BINARY      2
-#define OPCODE_CLOSE       8
-#define OPCODE_PING        9
-#define OPCODE_PONG       10
-#define HEADER_MAXLEN     16
-
-const char* opcode_string(uint16_t code)
-{
-  switch (code) {
-  case OPCODE_CONTINUE:
-      return "Continuation frame";
-  case OPCODE_TEXT:
-      return "Text frame";
-  case OPCODE_BINARY:
-      return "Binary frame";
-  case OPCODE_CLOSE:
-      return "Connection close";
-  case OPCODE_PING:
-      return "Ping";
-  case OPCODE_PONG:
-      return "Pong";
-  default:
-      return "Reserved (unspecified)";
-  }
-}
-
-namespace http {
-
-struct ws_header
-{
-  uint16_t bits;
-
-  bool is_final() const noexcept {
-    return (bits >> 7) & 1;
-  }
-  void set_final() noexcept {
-    bits |= 0x80;
-    assert(is_final() == true);
-  }
-  uint16_t payload() const noexcept {
-    return (bits >> 8) & 0x7f;
-  }
-  void set_payload(const size_t len)
-  {
-    uint16_t pbits = len;
-    if      (len > 65535) pbits = 127;
-    else if (len > 125)   pbits = 126;
-    bits &= 0x80ff;
-    bits |= (pbits & 0x7f) << 8;
-
-    if (is_ext())
-        *(uint16_t*) vla = __builtin_bswap16(len);
-    else if (is_ext2())
-        *(uint64_t*) vla = __builtin_bswap64(len);
-    assert(data_length() == len);
-  }
-
-  bool is_ext() const noexcept {
-    return payload() == 126;
-  }
-  bool is_ext2() const noexcept {
-    return payload() == 127;
-  }
-
-  bool is_masked() const noexcept {
-    return bits & 0x8000;
-  }
-  void set_masked() noexcept {
-    bits |= 0x8000;
-    uint32_t mask = OS::cycles_since_boot() & 0xffffffff;
-    memcpy(keymask(), &mask, sizeof(mask));
-  }
-
-  uint8_t opcode() const noexcept {
-    return bits & 0xf;
-  }
-  void set_opcode(uint8_t code) {
-    bits &= ~0xf;
-    bits |= code & 0xf;
-    assert(opcode() == code);
-  }
-
-  bool is_fail() const noexcept {
-    return false;
-  }
-
-  size_t mask_length() const noexcept {
-    return is_masked() ? 4 : 0;
-  }
-  char* keymask() noexcept {
-    return &vla[data_offset() - mask_length()];
-  }
-
-  size_t data_offset() const noexcept {
-    size_t len = mask_length();
-    if (is_ext2()) return len + 8;
-    if (is_ext())  return len + 2;
-    return len;
-  }
-  size_t data_length() const noexcept {
-    if (is_ext2())
-        return __builtin_bswap64(*(uint64_t*) vla) & 0xffffffff;
-    if (is_ext())
-        return __builtin_bswap16(*(uint16_t*) vla);
-    return payload();
-  }
-  const char* data() const noexcept {
-    return &vla[data_offset()];
-  }
-  char* data() noexcept {
-    return &vla[data_offset()];
-  }
-  void masking_algorithm()
-  {
-    char* ptr  = data();
-    const char* mask = keymask();
-    for (size_t i = 0; i < data_length(); i++)
-    {
-      ptr[i] = ptr[i] xor mask[i & 3];
-    }
-  }
-
-  size_t reported_length() const noexcept {
-    return sizeof(ws_header) + data_offset() + data_length();
-  }
-
-  char vla[0];
-} __attribute__((packed));
+namespace net {
 
 static inline std::string
 encode_hash(const std::string& key)
@@ -164,7 +35,7 @@ encode_hash(const std::string& key)
 }
 
 
-WebSocket_ptr WebSocket::upgrade(Request& req, Response_writer& writer)
+WebSocket_ptr WebSocket::upgrade(http::Request& req, http::Response_writer& writer)
 {
   // validate handshake
   auto view = req.header().value("Sec-WebSocket-Version");
@@ -196,7 +67,7 @@ WebSocket_ptr WebSocket::upgrade(Request& req, Response_writer& writer)
   return nullptr;
 }
 
-WebSocket_ptr WebSocket::upgrade(Error err, Response& res, Connection& conn, const std::string& key)
+WebSocket_ptr WebSocket::upgrade(http::Error err, http::Response& res, http::Connection& conn, const std::string& key)
 {
   if (err or res.status_code() != http::Switching_Protocols)
   {
@@ -230,15 +101,15 @@ std::vector<char> WebSocket::generate_key()
   return key;
 }
 
-Server::Request_handler WebSocket::create_request_handler(
+http::Server::Request_handler WebSocket::create_request_handler(
   Connect_handler on_connect, Accept_handler on_accept)
 {
-  auto handler = Server::Request_handler::make_packed(
+  auto handler = http::Server::Request_handler::make_packed(
     [
       on_connect{std::move(on_connect)},
       on_accept{std::move(on_accept)}
     ]
-    (Request_ptr req, Response_writer_ptr writer)
+    (http::Request_ptr req, http::Response_writer_ptr writer)
     {
       if (on_accept)
       {
@@ -259,15 +130,15 @@ Server::Request_handler WebSocket::create_request_handler(
   return handler;
 }
 
-Client::Response_handler WebSocket::create_response_handler(
+http::Client::Response_handler WebSocket::create_response_handler(
   Connect_handler on_connect, std::string key)
 {
-  auto handler = Client::Response_handler::make_packed(
+  auto handler = http::Client::Response_handler::make_packed(
     [
       on_connect{std::move(on_connect)},
       key{std::move(key)}
     ]
-    (Error err, Response_ptr res, Connection& conn)
+    (http::Error err, http::Response_ptr res, http::Connection& conn)
     {
       auto ws = WebSocket::upgrade(err, *res, conn, key);
 
@@ -333,14 +204,14 @@ void WebSocket::read_data(net::tcp::buffer_t buf, size_t len)
   }
 
   switch (hdr.opcode()) {
-  case OPCODE_TEXT:
-  case OPCODE_BINARY:
+  case op_code::TEXT:
+  case op_code::BINARY:
       /// .. call on_read
       if (on_read) {
           on_read(hdr.data(), hdr.data_length());
       }
       break;
-  case OPCODE_CLOSE:
+  case op_code::CLOSE:
       // they are angry with us :(
       if (hdr.data_length() >= 2) {
         // provide reason to user
@@ -354,18 +225,18 @@ void WebSocket::read_data(net::tcp::buffer_t buf, size_t len)
       // close it down
       this->close();
       return;
-  case OPCODE_PING:
-      write_opcode(OPCODE_PONG, hdr.data(), hdr.data_length());
+  case op_code::PING:
+      write_opcode(op_code::PONG, hdr.data(), hdr.data_length());
       return;
-  case OPCODE_PONG:
+  case op_code::PONG:
       return;
   default:
-      printf("Unknown opcode: %u\n", hdr.opcode());
+      printf("Unknown opcode: %u\n", static_cast<unsigned>(hdr.opcode()));
       break;
   }
 }
 
-static size_t make_header(char* dest, size_t len, uint8_t code, bool client)
+static size_t make_header(char* dest, size_t len, op_code code, bool client)
 {
   new (dest) ws_header;
   auto& hdr = *(ws_header*) dest;
@@ -374,13 +245,13 @@ static size_t make_header(char* dest, size_t len, uint8_t code, bool client)
   hdr.set_payload(len);
   hdr.set_opcode(code);
   if (client) {
-    hdr.set_masked();
+    hdr.set_masked(OS::cycles_since_boot() & 0xffffffff);
   }
   // header size + data offset
   return sizeof(ws_header) + hdr.data_offset();
 }
 
-void WebSocket::write(const char* buffer, size_t len, mode_t mode)
+void WebSocket::write(const char* buffer, size_t len, op_code code)
 {
   if (UNLIKELY(this->stream == nullptr)) {
     failure("write: Already closed");
@@ -390,11 +261,14 @@ void WebSocket::write(const char* buffer, size_t len, mode_t mode)
     failure("write: Connection not writable");
     return;
   }
-  uint8_t opcode = (mode == TEXT) ? OPCODE_TEXT : OPCODE_BINARY;
+
+  Expects(code == op_code::TEXT or code == op_code::BINARY
+    && "Write currently only supports TEXT or BINARY");
+
   // allocate header and data at the same time
-  auto buf = net::tcp::buffer_t(new uint8_t[HEADER_MAXLEN + len]);
+  auto buf = net::tcp::buffer_t(new uint8_t[WS_HEADER_MAXLEN + len]);
   // fill header
-  int  header_len = make_header((char*) buf.get(), len, opcode, clientside);
+  int  header_len = make_header((char*) buf.get(), len, code, clientside);
   // get data offset & fill in data into buffer
   char* data_ptr = (char*) buf.get() + header_len;
   memcpy(data_ptr, buffer, len);
@@ -409,7 +283,7 @@ void WebSocket::write(const char* buffer, size_t len, mode_t mode)
   /// send everything as shared buffer
   this->stream->write(buf, header_len + len);
 }
-void WebSocket::write(net::tcp::buffer_t buffer, size_t len, mode_t mode)
+void WebSocket::write(net::tcp::buffer_t buffer, size_t len, op_code code)
 {
   if (UNLIKELY(this->stream == nullptr)) {
     failure("write: Already closed");
@@ -423,22 +297,25 @@ void WebSocket::write(net::tcp::buffer_t buffer, size_t len, mode_t mode)
     failure("write: Client-side does not support sending shared buffers");
     return;
   }
-  uint8_t opcode = (mode == TEXT) ? OPCODE_TEXT : OPCODE_BINARY;
+
+  Expects(code == op_code::TEXT or code == op_code::BINARY
+    && "Write currently only supports TEXT or BINARY");
+
   /// write header
-  char header[HEADER_MAXLEN];
-  int  header_len = make_header(header, len, opcode, false);
-  assert(header_len < HEADER_MAXLEN);
+  char header[WS_HEADER_MAXLEN];
+  int  header_len = make_header(header, len, code, false);
+  assert(header_len < WS_HEADER_MAXLEN);
   this->stream->write(header, header_len);
   /// write shared buffer
   this->stream->write(buffer, len);
 }
-bool WebSocket::write_opcode(uint8_t code, const char* buffer, size_t datalen)
+bool WebSocket::write_opcode(op_code code, const char* buffer, size_t datalen)
 {
   if (UNLIKELY(stream == nullptr || stream->is_writable() == false)) {
     return false;
   }
   /// write header
-  char header[HEADER_MAXLEN];
+  char header[WS_HEADER_MAXLEN];
   int  header_len = make_header(header, datalen, code, clientside);
   this->stream->write(header, header_len);
   /// write buffer (if present)
@@ -478,7 +355,7 @@ void WebSocket::close()
 {
   /// send CLOSE message
   if (this->stream->is_writable())
-      this->write_opcode(OPCODE_CLOSE, nullptr, 0);
+      this->write_opcode(op_code::CLOSE, nullptr, 0);
   /// close and unset socket
   this->stream->close();
   this->reset();
@@ -534,4 +411,4 @@ const char* WebSocket::status_code(uint16_t code)
   }
 }
 
-} // http
+} // net
